scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.1"
description = ""
authors = []

[tool.poetry.scripts]
push_stats = "scripts.push_github_stats:main"
get_deps = "scripts.extract_path_deps:main"
update_poetry = "scripts.version_poetry_project:main"
update_next = "scripts.version_next_project:main"
get_tags = "scripts.get_commit_tags:main"
push_gh_stars = "scripts.push_github_stars:main"
push_gh_watches = "scripts.push_github_watches:main"
push_gh_forks = "scripts.push_github_forks:main"
push_pypi_downloads = "scripts.push_pypi_downloads:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
setuptools = "^69.0.2"
requests = "^2.31.0"
toml = "^0.10.2"
gitpython = "^3.1.43"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
ref-resolver = "^1.0.3"
jsonschema = "^4.22.0"
json-schema-for-humans = {version = "^1.0.2", optional = true}
jsonref = "^1.1.0"
pyyaml = "^6.0.1"
typer = "^0.9.0"

[tool.poetry.dependencies.eidolon-ai-sdk]
path = "../sdk"
develop = true
optional = true

[tool.poetry.group.dev.dependencies.eidolon-ai-usage-client]
path = "../usage-service/usage-client"
develop = true
optional = true

[tool.poetry.group.dev.dependencies.eidolon-ai-client]
path = "../client/python"
develop = true
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"

[tool.poetry.extras]
docbuilder = [ "eidolon-ai-sdk", "eidolon-ai-client", "eidolon-ai-usage-client", "json-schema-for-humans"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 121
